+++ src/command-name.ts
@@ -8,12 +8,14 @@ export enum CommandName {
 	Add = 'ionic.add',
 	SignUp = 'ionic.signUp',
 	Login = 'ionic.login',
+	Stop = 'ionic.stop',
 	Rebuild = 'ionic.rebuild',
 	Open = 'ionic.open',
 	SkipLogin = 'ionic.skipLogin',
 	Upgrade = 'ionic.upgrade',
 	ProjectsRefresh = 'ionic.projectRefresh',
-	ProjectSelect = 'ionic.projectSelect'
+	ProjectSelect = 'ionic.projectSelect',
+	BuildConfig = 'ionic.buildConfig'
 }
 
 export enum InternalCommand {