+++ src/extension.ts
@@ -3,7 +3,7 @@
 import * as vscode from 'vscode';
 import * as fs from 'fs';
 
-import { Context } from './context-variables';
+import { Context, VSCommand } from './context-variables';
 import { ionicLogin, ionicSignup } from './ionic-auth';
 import { ionicState, IonicTreeProvider } from './ionic-tree-provider';
 import { clearRefreshCache } from './process-packages';
@@ -16,6 +16,7 @@ import { handleError } from './error-handler';
 import { CommandName, InternalCommand } from './command-name';
 import { packageUpgrade } from './rules-package-upgrade';
 import { IonicProjectsreeProvider } from './ionic-projects-provider';
+import { buildConfiguration } from './build-configuration';
 
 
 let channel: vscode.OutputChannel = undefined;
@@ -151,14 +152,14 @@ function finishCommand(tip: Tip) {
 	});
 }
 
-function startCommand(tip: Tip, cmd: string) {
+function startCommand(tip: Tip, cmd: string, ionicProvider: IonicTreeProvider) {
 	if (tip.title) {
 		const message = tip.commandTitle ? tip.commandTitle : tip.title;
 		channel.appendLine(`[Ionic] ${message}...`);
 		let command = cmd;
 		if (command.includes(InternalCommand.cwd)) {
 			command = command.replace(InternalCommand.cwd, '');
-			channel.appendLine(`> Workspace: ${ionicState.workspace}`);		
+			channel.appendLine(`> Workspace: ${ionicState.workspace}`);
 		}
 		channel.appendLine(`> ${command}`);
 		channel.show();
@@ -192,6 +193,10 @@ export async function fixIssue(command: string | string[], rootPath: string, ion
 	// If the task is already running then cancel it
 	if (isRunning(tip)) {
 		await cancelRunning(tip);
+		if (tip.data == Context.stop) {
+			channel.show();
+			return; // User clicked stop
+		}
 	}
 
 	runningOperations.push(tip);
@@ -213,7 +218,8 @@ export async function fixIssue(command: string | string[], rootPath: string, ion
 				// Kill the process if the user cancels				
 				if (token.isCancellationRequested || tip.cancelRequested) {
 					tip.cancelRequested = false;
-					channel.appendLine(`Canceled operation "${tip.title}"`);
+					channel.appendLine(`[Ionic] Stopped "${tip.title}"`);
+					channel.show();
 					clearInterval(interval);
 					finishCommand(tip);
 					cancelObject.proc.kill();
@@ -232,7 +238,7 @@ export async function fixIssue(command: string | string[], rootPath: string, ion
 			if (Array.isArray(command)) {
 				try {
 					for (const cmd of command) {
-						startCommand(tip, cmd);
+						startCommand(tip, cmd, ionicProvider);
 						await run(rootPath, cmd, channel, cancelObject, tip.doViewEditor, tip.runPoints, progress, ionicProvider);
 
 					}
@@ -240,7 +246,7 @@ export async function fixIssue(command: string | string[], rootPath: string, ion
 					finishCommand(tip);
 				}
 			} else {
-				startCommand(tip, command);
+				startCommand(tip, command, ionicProvider);
 				const secondsTotal = estimateRunTime(command);
 				if (secondsTotal) {
 					increment = 100.0 / secondsTotal;
@@ -291,6 +297,12 @@ export function activate(context: vscode.ExtensionContext) {
 		}
 	});
 
+	vscode.commands.registerCommand(CommandName.Stop, async (recommendation: Recommendation) => {
+		recommendation.tip.data = Context.stop;
+		await fixIssue(undefined, context.extensionPath, ionicProvider, recommendation.tip);
+		recommendation.setContext(undefined);
+	});
+
 	vscode.commands.registerCommand(CommandName.SignUp, async () => {
 		await ionicSignup(context.extensionPath, context);
 		ionicProvider.refresh();
@@ -302,15 +314,22 @@ export function activate(context: vscode.ExtensionContext) {
 	});
 
 	vscode.commands.registerCommand(CommandName.Login, async () => {
-		await vscode.commands.executeCommand('setContext', Context.isLoggingIn, true);
+		await vscode.commands.executeCommand(VSCommand.setContext, Context.isLoggingIn, true);
 		await ionicLogin(context.extensionPath, context);
 		ionicProvider.refresh();
 	});
 
+	vscode.commands.registerCommand(CommandName.BuildConfig, async (r: Recommendation) => {
+		const config = await buildConfiguration(context.extensionPath, context, r.tip.actionArg(0));
+		if (!config) return;
+		r.tip.addActionArg(`--configuration=${config}`);
+		runAction(r, ionicProvider, rootPath);
+	});
+
 	vscode.commands.registerCommand(CommandName.SkipLogin, async () => {
 		ionicState.skipAuth = true;
-		await vscode.commands.executeCommand('setContext', Context.inspectedProject, false);
-		await vscode.commands.executeCommand('setContext', Context.isAnonymous, false);
+		await vscode.commands.executeCommand(VSCommand.setContext, Context.inspectedProject, false);
+		await vscode.commands.executeCommand(VSCommand.setContext, Context.isAnonymous, false);
 		ionicProvider.refresh();
 	});
 
@@ -335,8 +354,8 @@ export function activate(context: vscode.ExtensionContext) {
 	});
 
 	// User selected a project from the list (monorepo)
-	vscode.commands.registerCommand(CommandName.ProjectSelect, async (project: string) => {	
-		context.workspaceState.update('SelectedProject', project);	
+	vscode.commands.registerCommand(CommandName.ProjectSelect, async (project: string) => {
+		context.workspaceState.update('SelectedProject', project);
 		ionicProvider.selectProject(project);
 	});
 
@@ -344,31 +363,8 @@ export function activate(context: vscode.ExtensionContext) {
 		await fix(r.tip, rootPath, ionicProvider, context);
 	});
 
-	vscode.commands.registerCommand(CommandName.Run, async (tip: Tip) => {
-		tip.generateCommand();
-		tip.generateTitle();
-		if (tip.command) {
-			const info = tip.description ? tip.description : `${tip.title}: ${tip.message}`;
-			let command = tip.command;
-			if (tip.doRequestAppName) {
-				command = await requestAppName(tip);
-			}
-			if (tip.doDeviceSelection) {
-				const target = await selectDevice(tip.secondCommand as string, tip.data, tip);
-				if (!target) {
-					return;
-				}
-				command = (tip.command as string).replace(InternalCommand.target, target);
-			}
-			if (command) {
-				execute(tip);
-				fixIssue(command, rootPath, ionicProvider, tip);
-				return;
-			}
-		} else {
-			execute(tip);
-		}
-
+	vscode.commands.registerCommand(CommandName.Run, async (r: Recommendation) => {
+		runAction(r, ionicProvider, rootPath);
 	});
 
 	vscode.commands.registerCommand(CommandName.Link, async (tip: Tip) => {
@@ -376,6 +372,35 @@ export function activate(context: vscode.ExtensionContext) {
 	});
 }
 
+async function runAction(r: Recommendation, ionicProvider: IonicTreeProvider, rootPath: string) {
+	const tip = r.tip;
+	if (tip.stoppable) {
+		ionicProvider.refresh();
+	}
+	tip.generateCommand();
+	tip.generateTitle();
+	if (tip.command) {
+		const info = tip.description ? tip.description : `${tip.title}: ${tip.message}`;
+		let command = tip.command;
+		if (tip.doRequestAppName) {
+			command = await requestAppName(tip);
+		}
+		if (tip.doDeviceSelection) {
+			const target = await selectDevice(tip.secondCommand as string, tip.data, tip);
+			if (!target) {
+				return;
+			}
+			command = (tip.command as string).replace(InternalCommand.target, target);
+		}
+		if (command) {
+			execute(tip);
+			fixIssue(command, rootPath, ionicProvider, tip);
+			return;
+		}
+	} else {
+		execute(tip);
+	}
+}
 async function fix(tip: Tip, rootPath: string, ionicProvider: IonicTreeProvider, context: vscode.ExtensionContext): Promise<void> {
 	tip.generateCommand();
 	tip.generateTitle();